requirements.txt
aiogram==3.3.0
PyMuPDF==1.23.8
transformers==4.36.0
torch==2.1.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
ollama==0.1.7
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1